mv statics class to lib
