+++ bench/common/eval.py
@@ -1,0 +1,11 @@
+class Eval:
+    def __init__(self):
+        pass
+
+    def eval_type_acc(self, base, pre: dict):
+        """评估类型准确率."""
+        pass
+
+    def eval_content_acc(self, base, pre: str):
+        """评估内容准确率."""
+        pass
